requirements.txt
bentoml
torch
transformers
diffusers
accelerate
xformers
opencv-python
pillow
pydantic
